# Remove FxSwingUtils dependency on javafx-swing by implementing our own png image exporter