--- a/chartfx-chart/src/main/java/de/gsi/chart/utils/WriteFxImage.java
+++ b/chartfx-chart/src/main/java/de/gsi/chart/utils/WriteFxImage.java
@@ -1,32 +1,32 @@
 package de.gsi.chart.utils;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.nio.ByteBuffer;
-import java.nio.channels.FileChannel;
+import java.nio.file.Files;
+import java.util.Map;
 import java.util.zip.CRC32;
 import java.util.zip.Deflater;
 
 import javafx.scene.image.Image;
 import javafx.scene.image.PixelReader;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import de.gsi.dataset.utils.ArrayCache;
 
 /**
  * Writes a JavaFx Image into a ByteBuffer or file
- * TODO: possible improvements
+ * 
+ * possible improvements:
  * - use multiple IDAT chunks to use fixed buffer size and limit memory footprint
- * - implement filtering for smaller file sizes
+ * - implement filtering of lines before compression for smaller file sizes
  * - Optionally add tEXT chunks for metadata (EXIF)
  *
  * @author Alexander Krimm
  */
-public class WriteFxImage {
-    private static final Logger LOGGER = LoggerFactory.getLogger(WriteFxImage.class);
-    private static final int HEADER_SIZE = 8 + (12 + 13) + 12 + 12; // size of all the headers and other Metadata
-    private static final int BUFFER_MARGIN = 200; // how many additional bytes to allocate for the buffers
+public final class WriteFxImage {
+    protected static final int HEADER_SIZE = 8 + (12 + 13) + 12 + 12; // size of all the headers and other Metadata
+    protected static final String INTERNAL_ARRAY_CACHE_NAME = "WriteFxImage-internalArray";
 
     /**
      * private constructor for static utility class
@@ -35,144 +35,198 @@ public class WriteFxImage {
     }
 
     /**
-     * Encodes a JavaFx image as an RGB png image
+     * Encodes a JavaFx image as an RGB png image with maximum (lossless) compression 
      *
-     * @param image an javafx.scene.image.Image
+     * @param image The input image to be encoded
      * @return a byte buffer with the encoded image
-     * @see "https://tools.ietf.org/html/rfc2083"
+     * @see <a href="https://tools.ietf.org/html/rfc2083">rfc2083</a>
      */
     public static ByteBuffer encode(final Image image) {
-        return encode(image, null);
+        return encode(image, null, true, Deflater.BEST_COMPRESSION, null);
     }
 
     /**
      * Encodes a JavaFx image as an RGB png image.
      * If you pass in a ByteBuffer to use, please make sure that it has enough capacity to fit the encoded image or
-     * handle errors (IndexOutOfBoundsException) accordingly.
+     * handle errors (IndexOutOfBoundsException) accordingly. If you want to be on the safe side, use
+     * {@link #getCompressedSizeBound(int, int, boolean) getCompressedSizeBound(width, height, alpha)} to get an upper bound for the output size.
      * 
-     * @param image an javafx.scene.image.Image
+     * @param image The input image to be encoded
      * @param byteBuffer optional byte buffer to store the output in, pass null to return a new one.
+     * @param alpha wheter to include alpha information in the image
+     * @param compressionLevel {@link Deflater#BEST_COMPRESSION} (9) to {@link Deflater#BEST_SPEED} (0)
+     * @param metaInfo an optional map which will be filled with debugging information like compression efficiency
      * @return a byte buffer with the encoded image
      * @see "https://tools.ietf.org/html/rfc2083"
      */
-    public static ByteBuffer encode(final Image image, final ByteBuffer byteBuffer) {
-        if (byteBuffer != null) {
-            byteBuffer.clear();
-        }
-        int w = (int) image.getWidth();
-        int h = (int) image.getHeight();
-        // conservative allocation, assume zero compression
-        final ByteBuffer outputByteBuffer = byteBuffer == null ? ByteBuffer.allocate((w + 1) * h * 3 + HEADER_SIZE + BUFFER_MARGIN) : byteBuffer;
+    public static ByteBuffer encode(final Image image, final ByteBuffer byteBuffer,final boolean alpha,
+            final int compressionLevel, final Map<String, Object> metaInfo) {
+        // get meta info
+        final int w = (int) image.getWidth();
+        final int h = (int) image.getHeight();
+        // allocate output buffer if necessary. conservative allocation, assume upper bound for deflate algorithm
+        final ByteBuffer outputByteBuffer = byteBuffer != null ? byteBuffer : ByteBuffer.allocate(getCompressedSizeBound(w, h, alpha));
+        // get necessary helper classes
         final PixelReader pr = image.getPixelReader();
         CRC32 crc = new CRC32();
-
-        // File Signature - "\211PNG\r\n\032\n" - 8950 4e47 0d0a 1a0a
-        outputByteBuffer.put(new byte[] { 0x89 - 0xff - 1, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a });
-
-        // IHDR
-        write(13, outputByteBuffer); // length of IHDR chunk
-        crc.reset(); // checksum starts after length field
-        write("IHDR".getBytes(), outputByteBuffer, crc);
-        write(w, outputByteBuffer, crc); // with 4 bytes
-        write(h, outputByteBuffer, crc); // height 4 bytes
-        // Bit depth:          1 byte  1*,2*,4*,8,16° (*:indexed only, °: not indexed)
-        // Color type:         1 byte  0 (grayscale), 2 (RGB), 3 (Indexed), 4 (grayscale+alpha), and 6 (RGBA).
-        // Compression method: 1 byte  0 (deflate/inflate compression with a 32K sliding window)
-        // Filter method:      1 byte  0 (adaptive filtering with five basic filter types)
-        // Interlace method:   1 byte  0 (no interlace) or 1 (Adam7 interlace)
-        write(new byte[] { 8, 2, 0, 0, 0 }, outputByteBuffer, crc); // RGB Mode
-        write((int) crc.getValue(), outputByteBuffer);
-
-        int sizePos = outputByteBuffer.position();
-        write(0, outputByteBuffer); // zero size, will later be overwritten when we know the size
-        crc.reset();
-        write("IDAT".getBytes(), outputByteBuffer, crc);
-        // compress image data
-        final Deflater compressor = new Deflater(Deflater.BEST_COMPRESSION);
-        final ByteBuffer uncompressedImageData = ByteBuffer.allocate((w + 1) * h * 3);
-        compressor.setInput(uncompressedImageData);
-        for (int y = 0; y < h; y++) {
-            uncompressedImageData.put((byte) 0); // line filtering: 0: None 1: Sub 2: Up 3: Average 4: Paeth
-            for (int x = 0; x < w; x++) {
-                int pixel = pr.getArgb(x, y);
-                uncompressedImageData.put((byte) ((pixel >> 16) & 0xff));
-                uncompressedImageData.put((byte) ((pixel >> 8) & 0xff));
-                uncompressedImageData.put((byte) (pixel & 0xff));
-            }
-        }
-        // Does not work because rgb pixel format is not writable even if isWritable says so -.-
-        // final WritablePixelFormat<ByteBuffer> pixelformat = (WritablePixelFormat<ByteBuffer>) PixelFormat
-        //         .getByteRgbInstance();
-        // uncompressedImageData.put((byte) 0); // line filtering: 0: None 1: Sub 2: Up 3: Average 4: Paeth
-        // pr.getPixels(0, 0, w, h, pixelformat, uncompressedImageData, w + 1); // stride of w+1 puts line filtering to 0(None)
-        uncompressedImageData.flip();
-        compressor.finish();
-        compressor.deflate(outputByteBuffer, Deflater.FULL_FLUSH);
-        // update crc
-        crc.update(outputByteBuffer.array(), sizePos + 8, compressor.getTotalOut());
-        outputByteBuffer.putInt(sizePos, compressor.getTotalOut());
-        write((int) crc.getValue(), outputByteBuffer);
-
-        // IEND chunk
-        write(0, outputByteBuffer);
-        crc.reset();
-        write("IEND".getBytes(), outputByteBuffer, crc);
-        write((int) crc.getValue(), outputByteBuffer);
-
+        final Deflater compressor = new Deflater(compressionLevel);
+        // actual PNG encoding
+        writeImageHeader(w, h, alpha, outputByteBuffer, crc);
+        writeImageData(pr, w, h, alpha, compressor, outputByteBuffer, crc);
+        writeImageFooter(outputByteBuffer, crc);
+        // prepare to return outputBuffer
         outputByteBuffer.flip();
-
         // DEGBUG output
-        if (LOGGER.isDebugEnabled()) {
-            final float compression = (100.0f * compressor.getBytesWritten()) / compressor.getBytesRead();
-            LOGGER.atDebug() //
-                    .addArgument(w)
-                    .addArgument(h) //
-                    .addArgument(compressor.getBytesRead()) //
-                    .addArgument(compressor.getBytesWritten()) //
-                    .addArgument(compression) //
-                    .addArgument(outputByteBuffer.limit() - outputByteBuffer.position()) //
-                    .log("Convertet Image({}x{}) to png: IN: {} bytes, OUT: {} bytes, compression: {}%, total file size: {} bytes");
+        if (metaInfo != null) {
+            final int bytesPerPixel = alpha ? 4 : 3;
+            metaInfo.put("bufferSize", outputByteBuffer.capacity());
+            metaInfo.put("outputSizeBound", getCompressedSizeBound(w, h, alpha));
+            metaInfo.put("compression", ((double) compressor.getBytesWritten()) / (w * h * bytesPerPixel));
+            metaInfo.put("width", w);
+            metaInfo.put("height", h);
+            metaInfo.put("colorMode", alpha ? "rgba" : "rgb");
+            metaInfo.put("compressionLevel", compressionLevel);
+            metaInfo.put("outputSize", outputByteBuffer.limit() - outputByteBuffer.position());
         }
-
         return outputByteBuffer;
     }
 
     /**
-     * Saves the given Image as a png file.
+     * Saves the given image as a png file.
      *
      * @param image The image to save
      * @param file The filename to save the image to.
      * @throws IOException if the file cannot be written
      */
     public static void savePng(Image image, File file) throws IOException {
-        try (FileOutputStream os = new FileOutputStream(file); FileChannel wChannel = os.getChannel()) {
+        try (OutputStream os = Files.newOutputStream(file.toPath())) {
             ByteBuffer buffer = WriteFxImage.encode(image);
-            wChannel.write(buffer);
+            os.write(buffer.array(), 0, buffer.limit());
         }
+    }
+
+    /**
+     * Returns the conservative upper bound for the compressed image size.
+     * 
+     * @param width Image width
+     * @param height Image height
+     * @param alpha Alpha enabled
+     * @return the upper bound for the size of the resulting png in bytes
+     * @see <a href="https://github.com/madler/zlib/blob/master/deflate.c#L659-L661">zlib sourcefor deflate upper bound</a>
+     */
+    public static int getCompressedSizeBound(final int width, final int height, final boolean alpha) {
+        final int bytesPerPixel = alpha ? 4 : 3;
+        int uncompressedSize = (width) * height * bytesPerPixel + height + HEADER_SIZE;
+        int compressedSize = uncompressedSize + ((uncompressedSize + 7) >> 3) + ((uncompressedSize + 63) >> 6) + 5;
+        return compressedSize + HEADER_SIZE;
+    }
+
+    /**
+     * Writes the PNG file header and IHDR meta-information block
+     * 
+     * @param width The with of the image
+     * @param height The height of the image
+     * @param alpha whether to support alpha
+     * @param outputByteBuffer the buffer to write into
+     * @param crc the checksum calculator
+     */
+    private static void writeImageHeader(int width, int height, boolean alpha, ByteBuffer outputByteBuffer, CRC32 crc) {
+        // File Signature - "\211PNG\r\n\032\n" - 8950 4e47 0d0a 1a0a
+        outputByteBuffer.put(new byte[] { (byte) 0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a });
+        // IHDR
+        outputByteBuffer.putInt(13); // length of IHDR chunk
+        crc.reset(); // checksum starts after length field
+        write("IHDR".getBytes(), outputByteBuffer, crc);
+        write(width, outputByteBuffer, crc); // with 4 bytes
+        write(height, outputByteBuffer, crc); // height 4 bytes
+        // Bit depth:          1 byte  1*,2*,4*,8,16° (*:indexed only, °: not indexed)
+        // Color type:         1 byte  0 (grayscale), 2 (RGB), 3 (Indexed), 4 (grayscale+alpha), and 6 (RGBA).
+        // Compression method: 1 byte  0 (deflate/inflate compression with a 32K sliding window)
+        // Filter method:      1 byte  0 (adaptive filtering with five basic filter types)
+        // Interlace method:   1 byte  0 (no interlace) or 1 (Adam7 interlace)
+        write(alpha ? new byte[] { 8, 6, 0, 0, 0 } : new byte[] { 8, 2, 0, 0, 0 }, outputByteBuffer, crc); // RGB(A) Mode
+        outputByteBuffer.putInt((int) crc.getValue());
+    }
+
+    /**
+     * Writes the IDAT chunk of a png file, which contains the compressed image data to the supplied buffer.
+     * 
+     * @param pr PixelWriter
+     * @param w with of the image
+     * @param h height of the image
+     * @param alpha whether to include alpha information
+     * @param outputByteBuffer the byte buffer to write to
+     * @param crc checksum calculator
+     */
+    private static void writeImageData(PixelReader pr, int w, int h, boolean alpha, Deflater compressor,
+            ByteBuffer outputByteBuffer, CRC32 crc) {
+        // get raw image data
+        final int bytesPerPixel = alpha ? 4 : 3;
+        final int rawDataSize = w * h * bytesPerPixel + h; // image dimensions times bytesPerPixel + line filtering flag
+        final byte[] uncompressedImageData = ArrayCache.getCachedByteArray(INTERNAL_ARRAY_CACHE_NAME, rawDataSize);
+        int i = 0;
+        if (alpha) {
+            for (int y = 0; y < h; y++) {
+                uncompressedImageData[i++] = 0; // LineFiltering: 0: None 1: Sub 2: Up 3: Average 4: Paeth
+                for (int x = 0; x < w; x++) {
+                    int pixel = pr.getArgb(x, y);
+                    uncompressedImageData[i++] = (byte) ((pixel >> 16) & 0xff); //red
+                    uncompressedImageData[i++] = (byte) ((pixel >> 8) & 0xff); // green
+                    uncompressedImageData[i++] = (byte) (pixel & 0xff); // blue
+                    uncompressedImageData[i++] = (byte) (pixel >> 24 & 0xff); // alpha
+                }
+            }
+        } else {
+            for (int y = 0; y < h; y++) {
+                uncompressedImageData[i++] = 0; // LineFiltering: 0: None 1: Sub 2: Up 3: Average 4: Paeth
+                for (int x = 0; x < w; x++) {
+                    int pixel = pr.getArgb(x, y);
+                    uncompressedImageData[i++] = (byte) ((pixel >> 16) & 0xff); //red
+                    uncompressedImageData[i++] = (byte) ((pixel >> 8) & 0xff); // green
+                    uncompressedImageData[i++] = (byte) (pixel & 0xff); // blue
+                }
+            }
+        }
+        // write compressed image data to IDAT block
+        compressor.setInput(uncompressedImageData);
+        outputByteBuffer.mark();
+        outputByteBuffer.putInt(0); // zero size, will later be overwritten when we know the size
+        outputByteBuffer.put("IDAT".getBytes());
+        compressor.finish();
+        compressor.deflate(outputByteBuffer, Deflater.FULL_FLUSH);
+        outputByteBuffer.limit(outputByteBuffer.position());
+        outputByteBuffer.reset();
+        outputByteBuffer.putInt(compressor.getTotalOut());
+        crc.reset();
+        crc.update(outputByteBuffer);
+        outputByteBuffer.limit(outputByteBuffer.capacity());
+        outputByteBuffer.putInt((int) crc.getValue());
+        ArrayCache.release(INTERNAL_ARRAY_CACHE_NAME, uncompressedImageData);
+    }
+
+    /**
+     * Writes the end marker for the png file format
+     * 
+     * @param outputByteBuffer the buffer to write into
+     * @param crc The checksum calculator
+     */
+    private static void writeImageFooter(ByteBuffer outputByteBuffer, CRC32 crc) {
+        outputByteBuffer.putInt(0);
+        crc.reset();
+        write("IEND".getBytes(), outputByteBuffer, crc);
+        outputByteBuffer.putInt((int) crc.getValue());
     }
 
     /**
      * Writes a byte array to the buffer and updates the checksum.
      *
-     * @param b
-     * @param buffer
-     * @param crc
+     * @param b byteArray to put into buffer
+     * @param buffer output buffer
+     * @param crc checksum calculator
      */
-    private static void write(byte b[], ByteBuffer buffer, CRC32 crc) {
+    private static void write(byte[] b, ByteBuffer buffer, CRC32 crc) {
         buffer.put(b);
         crc.update(b);
-    }
-
-    /**
-     * Writes an integer value to the buffer.
-     *
-     * @param i integer value to write
-     * @param buffer the buffer to write to
-     */
-    private static void write(int i, ByteBuffer buffer) {
-        byte b[] = { (byte) ((i >> 24) & 0xff), (byte) ((i >> 16) & 0xff), (byte) ((i >> 8) & 0xff),
-            (byte) (i & 0xff) };
-        buffer.put(b);
     }
 
     /**
@@ -183,8 +237,8 @@ public class WriteFxImage {
      * @param crc the checksum to be updated
      */
     private static void write(int i, ByteBuffer buffer, CRC32 crc) {
-        byte b[] = { (byte) ((i >> 24) & 0xff), (byte) ((i >> 16) & 0xff), (byte) ((i >> 8) & 0xff),
-            (byte) (i & 0xff) };
+        byte[] b = { (byte) ((i >> 24) & 0xff), (byte) ((i >> 16) & 0xff), (byte) ((i >> 8) & 0xff),
+                (byte) (i & 0xff) };
         write(b, buffer, crc);
     }
 }