… chartfx-chart/pom.xml
… chartfx-chart/src/test/java/de/gsi/chart/plugins/ScreenshotTest.java
… chartfx-chart/src/test/java/de/gsi/chart/utils/WriteFxImageBenchmark.java
… chartfx-chart/src/test/java/de/gsi/chart/utils/WriteFxImageTests.java