Implements exporting a JavaFx Image to png either directly to file or to a ByteBuffer, uses it to replace all occurences of FXSwingUtils and remove the javafx-swing dependency.