@RalphSteinhagen 2020-04-03
Verry good PR that significantly improves the chart-fx library as well as JavaFX as a whole!!
Good job @wirew0rm :+1: 

Fixed minor PMD violations, formatting and added a compression-level benchmark. In response to that, changed the default setting from `Deflater.BEST_COMPRESSION` to `Deflater.BEST_SPEED` since the compression gain was indeed insignificant w.r.t. to the CPU load increase as suggested by @ennerf

In any case, `Deflater.BEST_COMPRESSION` can still be achieved using the `encode(..)` method directly.

Performance results for reference:
```
Test compression level performance with image with noise data (difficult to compress)
with alpha channel
FxImage: size 333x777 	 compression: 100.0% 	  6.10 ms/image    rgba keepBuffer compressionLevel: 0
FxImage: size 333x777 	 compression:  82.6% 	 49.46 ms/image    rgba keepBuffer compressionLevel: 1
FxImage: size 333x777 	 compression:  82.3% 	 55.80 ms/image    rgba keepBuffer compressionLevel: 2
FxImage: size 333x777 	 compression:  82.2% 	 58.12 ms/image    rgba keepBuffer compressionLevel: 3
FxImage: size 333x777 	 compression:  80.9% 	 69.82 ms/image    rgba keepBuffer compressionLevel: 4
FxImage: size 333x777 	 compression:  80.8% 	 71.08 ms/image    rgba keepBuffer compressionLevel: 5
FxImage: size 333x777 	 compression:  80.8% 	 71.96 ms/image    rgba keepBuffer compressionLevel: 6
FxImage: size 333x777 	 compression:  80.8% 	 72.28 ms/image    rgba keepBuffer compressionLevel: 7
FxImage: size 333x777 	 compression:  80.8% 	 72.76 ms/image    rgba keepBuffer compressionLevel: 8
FxImage: size 333x777 	 compression:  80.8% 	 72.76 ms/image    rgba keepBuffer compressionLevel: 9
without alpha channel
FxImage: size 333x777 	 compression: 100.0% 	  6.40 ms/image    rgb  keepBuffer compressionLevel: 0
FxImage: size 333x777 	 compression:  94.0% 	 35.90 ms/image    rgb  keepBuffer compressionLevel: 1
FxImage: size 333x777 	 compression:  94.0% 	 35.92 ms/image    rgb  keepBuffer compressionLevel: 2
FxImage: size 333x777 	 compression:  94.0% 	 36.08 ms/image    rgb  keepBuffer compressionLevel: 3
FxImage: size 333x777 	 compression:  93.5% 	 41.02 ms/image    rgb  keepBuffer compressionLevel: 4
FxImage: size 333x777 	 compression:  93.5% 	 39.46 ms/image    rgb  keepBuffer compressionLevel: 5
FxImage: size 333x777 	 compression:  93.5% 	 40.38 ms/image    rgb  keepBuffer compressionLevel: 6
FxImage: size 333x777 	 compression:  93.5% 	 40.20 ms/image    rgb  keepBuffer compressionLevel: 7
FxImage: size 333x777 	 compression:  93.5% 	 40.54 ms/image    rgb  keepBuffer compressionLevel: 8
FxImage: size 333x777 	 compression:  93.5% 	 40.70 ms/image    rgb  keepBuffer compressionLevel: 9
Test compression level performance with image with simple shapes (easy to compress)
with alpha channel
FxImage: size 777x333 	 compression: 100.0% 	  6.30 ms/image    rgba keepBuffer compressionLevel: 0
FxImage: size 777x333 	 compression:   0.9% 	  9.10 ms/image    rgba keepBuffer compressionLevel: 1
FxImage: size 777x333 	 compression:   0.9% 	  8.54 ms/image    rgba keepBuffer compressionLevel: 2
FxImage: size 777x333 	 compression:   0.9% 	  8.22 ms/image    rgba keepBuffer compressionLevel: 3
FxImage: size 777x333 	 compression:   0.5% 	 14.22 ms/image    rgba keepBuffer compressionLevel: 4
FxImage: size 777x333 	 compression:   0.5% 	 14.12 ms/image    rgba keepBuffer compressionLevel: 5
FxImage: size 777x333 	 compression:   0.5% 	 14.42 ms/image    rgba keepBuffer compressionLevel: 6
FxImage: size 777x333 	 compression:   0.5% 	 14.68 ms/image    rgba keepBuffer compressionLevel: 7
FxImage: size 777x333 	 compression:   0.5% 	 16.54 ms/image    rgba keepBuffer compressionLevel: 8
FxImage: size 777x333 	 compression:   0.5% 	 21.86 ms/image    rgba keepBuffer compressionLevel: 9
without alpha channel
FxImage: size 777x333 	 compression: 100.0% 	  6.40 ms/image    rgb  keepBuffer compressionLevel: 0
FxImage: size 777x333 	 compression:   1.1% 	  8.02 ms/image    rgb  keepBuffer compressionLevel: 1
FxImage: size 777x333 	 compression:   1.1% 	  7.94 ms/image    rgb  keepBuffer compressionLevel: 2
FxImage: size 777x333 	 compression:   1.1% 	  8.00 ms/image    rgb  keepBuffer compressionLevel: 3
FxImage: size 777x333 	 compression:   0.7% 	 12.34 ms/image    rgb  keepBuffer compressionLevel: 4
FxImage: size 777x333 	 compression:   0.7% 	 12.80 ms/image    rgb  keepBuffer compressionLevel: 5
FxImage: size 777x333 	 compression:   0.6% 	 13.02 ms/image    rgb  keepBuffer compressionLevel: 6
FxImage: size 777x333 	 compression:   0.6% 	 13.40 ms/image    rgb  keepBuffer compressionLevel: 7
FxImage: size 777x333 	 compression:   0.6% 	 16.18 ms/image    rgb  keepBuffer compressionLevel: 8
FxImage: size 777x333 	 compression:   0.6% 	 19.28 ms/image    rgb  keepBuffer compressionLevel: 9
```