@ennerf 2020-04-03
Yes, thanks for adding a custom png export! This gets rid of a very annoying dependency.